refactor(*): Extract MetricsVariableFilterEngine
